(wip) remove dynamic layer
