+++ CMakeLists.txt
@@ -35,12 +35,10 @@ configure_file(cmake/spark_dsg_version.h.in include/spark_dsg_version.h)
 add_library(
   ${PROJECT_NAME}
   src/adjacency_matrix.cpp
-  src/base_layer.cpp
   src/bounding_box_extraction.cpp
   src/bounding_box.cpp
   src/color.cpp
   src/colormaps.cpp
-  src/dynamic_scene_graph_layer.cpp
   src/dynamic_scene_graph.cpp
   src/edge_attributes.cpp
   src/edge_container.cpp
