+++ tests/CMakeLists.txt
@@ -8,7 +8,6 @@ add_executable(
   utest_bounding_box_extraction.cpp
   utest_bounding_box.cpp
   utest_color.cpp
-  utest_dynamic_scene_graph_layer.cpp
   utest_dynamic_scene_graph.cpp
   utest_edge_container.cpp
   utest_graph_utilities_layer.cpp
